Exchanging atexit mechanism for resultcallback of click
Atexit mechanism is called every time even when exception is called,
this could result easily in inconsistent Store's state.

Click supports resultcallback which is here used for automatically saving
Store state. This callback is not called when exception is raised.

This mechanism might still lead to inconsistent Store's state as giTrack
is depending on external services, but it seems better approach to
lost possibly corrupted state (there is Exception raised) and require
the user to repeat the action, over committing the corrupted state.